requirements.txt
Flask
Werkzeug
mysql-connector
dotenv
colorama